+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add support for manual HTTP framing headers ([#551](https://github.com/fastly/Viceroy/pull/551))
+
 ## 0.15.0 (2025-10-27)
 
 - fix: don't throw error when exit code is 0 ([#537](https://github.com/fastly/Viceroy/pull/537))